Factor out component-specific CSS into modules
- This begins the process of paring down our custom CSS and
  making it more modular.
- Making adjustments to CSS modules that don't impact the USWDS
  theme will result in faster rebuilds now.

See https://remix.run/docs/en/main/guides/styling#css-modules
