+++ .gitignore
@@ -14,7 +14,8 @@ sam.json
 sam.yaml
 
 # Generated by sass
-/app/css
+/app/**/*.css
+/app/**/*.css.map
 
 .env
 .eslintcache
