+++ app/components/nested-checkboxes/nested-checkboxes.module.scss
@@ -1,0 +1,24 @@
+.node {
+  & {
+    list-style: url('~/../node_modules/nasawds/src/img/usa-icons/chevron_right.svg');
+  }
+
+  &[aria-expanded='true'] {
+    list-style: url('~/../node_modules/nasawds/src/img/usa-icons/expand_more.svg');
+  }
+
+  & [data-testid='checkbox'] {
+    position: relative;
+    top: -0.25rem;
+    display: inline-block;
+  }
+
+  & label {
+    margin-top: 0;
+    margin-bottom: 0.5rem;
+  }
+}
+
+.leaf {
+  list-style: none;
+}
