+++ app/components/tabs/tabs.module.scss
@@ -1,0 +1,46 @@
+.tabs {
+  & [role='tablist'] {
+    & {
+      display: flex;
+      flex-wrap: wrap;
+      border-bottom: 1px solid #dee2e6;
+      position: relative;
+    }
+
+    & button {
+      & {
+        border-width: 1px;
+        border-style: solid;
+        border-color: transparent;
+        border-top-left-radius: 0;
+        border-top-right-radius: 0;
+        display: block;
+        padding: 0.5rem 1rem;
+        cursor: pointer;
+        margin: 0;
+        outline: none;
+        background-color: #fff;
+        transition: border 0.2s ease-in-out 0s;
+      }
+
+      &.active,
+      &:focus,
+      &:hover {
+        border-width: 1px;
+        border-color: #e9ecef #e9ecef #b41c1c;
+      }
+
+      &.active {
+        color: #495057;
+        background-color: #f8f8f8;
+      }
+    }
+  }
+
+  & [role='tabpanel'] {
+    text-align: left;
+    padding: 1rem;
+    background-color: #f8f8f8;
+    box-shadow: 1px 1px 2px rgb(204 204 204 / 75%);
+  }
+}
