+++ package.json
@@ -9,17 +9,17 @@
   "license": "Apache-2.0",
   "scripts": {
     "build:remix": "remix build",
-    "build:sass": "sass -Inode_modules/nasawds/src/theme -Inode_modules/@uswds -Inode_modules/@uswds/uswds/packages app/sass:app/css",
+    "build:sass": "sass -Inode_modules/nasawds/src/theme -Inode_modules/@uswds -Inode_modules/@uswds/uswds/packages app:app",
     "build:esbuild": "node esbuild.mjs",
     "build:website": "run-s build:sass build:remix",
     "build": "run-p build:website build:esbuild",
     "dev:remix": "remix dev --manual -c \"arc sandbox -e testing --host localhost\"",
-    "dev:sass": "sass --watch -Inode_modules/nasawds/src/theme -Inode_modules/@uswds -Inode_modules/@uswds/uswds/packages app/sass:app/css",
+    "dev:sass": "sass --watch -Inode_modules/nasawds/src/theme -Inode_modules/@uswds -Inode_modules/@uswds/uswds/packages app:app",
     "dev:esbuild": "node esbuild.mjs --dev",
     "dev": "run-p \"dev:*\"",
     "prepare": "husky install",
     "deploy": "arc deploy --no-hydrate --prune --production",
-    "clean": "rimraf --glob build/static app/css build/server/index.* build/server/metafile.* build/email-incoming build/table-streams sam.json sam.yaml .cache",
+    "clean": "rimraf --glob build/static app/css build/server/index.* build/server/metafile.* \"app/**/*.css\" \"app/**/*.css.map\" build/email-incoming build/table-streams sam.json sam.yaml .cache",
     "test": "jest",
     "test-coverage": "jest --coverage",
     "typecheck": "tsc"
@@ -33,6 +33,7 @@
     "@octokit/rest": "^20.0.1",
     "@opensearch-project/opensearch": "^2.2.0",
     "@remix-run/architect": "^1.19.3",
+    "@remix-run/css-bundle": "^1.19.3",
     "@remix-run/node": "^1.19.3",
     "@remix-run/react": "^1.19.3",
     "@trussworks/react-uswds": "^5.1.1",
